move flurry power
